+++ config.js
@@ -1,3 +1,4 @@
 module.exports = {
     valid_cache_domains: ['google.com', 'gstatic.com'],
+    recaptcha_version_check: 1,  // hours
 };